Fixed Comsk to run both direction and fixed
Physically added a new effekt relay to be able to control both comsk drives correctly making the product convayors able to run both directions, this has also been implemented in the code.

Changed what dimension can be set if no robot is with the system when creating a new item and what unit the dimensions is.

Removed add and delete buttons when editing item and pattern since there is allready add new and remove buttons

Cleaned database so mostly the only system avaible is 708 which is the one running on the CP6000